feat: refactor Navigation Component
