+++ .storybook/preview.js
@@ -4,6 +4,7 @@ import { customViewports } from './viewports';
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
   controls: {
+    hideNoControlsWarning: true,
     matchers: {
       color: /(background|color)$/i,
       date: /Date$/,
