--- src/components/navigation/navigation.stories.tsx
@@ -1,26 +1,0 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react';
-import React from 'react';
-import { Navigation } from './navigation';
-
-export default {
-  title: 'Components/Navigation',
-  component: Navigation,
-  parameters: {
-    backgrounds: { default: 'surface' },
-    layout: 'centered',
-  },
-} as ComponentMeta<typeof Navigation>;
-
-const Template: ComponentStory<typeof Navigation> = (args) => <Navigation {...args} />;
-
-export const Default = Template.bind({});
-Default.args = {
-  alt: 'Robert Vogt',
-  src: 'https://media.licdn.com/dms/image/D4E03AQEXHsHgH4BwJg/profile-displayphoto-shrink_800_800/0/1666815812197?e=2147483647&v=beta&t=Vx6xecdYFjUt3UTCmKdh2U-iHvY0bS-fcxlp_LKbxYw',
-};
-Default.parameters = {
-  design: {
-    type: 'figma',
-    url: 'https://www.figma.com/file/nsXR2h0KwciWpuwKRD58FX/Mumble?node-id=447%3A780&t=dW7d9mR1MQqmSsA6-4',
-  },
-};
