+++ src/index.ts
@@ -4,4 +4,4 @@ export * from './components/text/public-api';
 export * from './components/icons/public-api';
 export * from './components/logos/public-api';
 export * from './components/profile-picture/public-api';
-export * from './components/navigation/public-api';
+export * from './layout/navigation/public-api';
